language filesystem, howto use: at installation-time only en.lang exists then you can go to admin-languages and install languages from the masterfile
later in the developmentprocess one will turn the generated files into the database, but now it is fast enough and:
the smaller lang-files, e.g. "de.lang" can operate as intermediate file for a translator.
If she/he has finished translation she/he will give it to a developer. he will merge the intermediate file to the master fil
e then commit.
okay?

so press the generate from file-button in the admin-languages section and all languagefiles are created.

the backmerging into the large languagesfile is not written yet


git-svn-id: http://svn.ilias.de/svn/ilias/trunk@34 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5